requirements.txt
aiosqlite~=0.20
bittensor==7.3.0
diskcache~=5.6
fastapi~=0.99
httpx~=0.27
imagehash~=4.3
opencv-python~=4.10
pydantic~=1.10
python-dotenv~=1.0
scikit-learn~=1.5
redis~=5.0
sqlite-web==0.6.3
typer==0.10.0
xgboost==2.0.3
scikit-learn
asyncpg==0.29.0
# numpy==1.26.4

